* Moved relevant files to a framework folder.
