+++ qaoa_training_pipeline/framework/problem_params_provider.py
@@ -1,4 +1,4 @@
-#
+﻿#
 #
 # (C) Copyright IBM 2024.
 #
@@ -22,8 +22,8 @@ from abc import abstractmethod
 from qiskit import QuantumCircuit
 from qiskit.quantum_info import SparsePauliOp
 
-from qaoa_training_pipeline.qaoa_training_pipeline.params_provider import ParamsProvider
-from qaoa_training_pipeline.training.param_result import ParamResult
+from qaoa_training_pipeline.framework.params_provider import ParamsProvider
+from qaoa_training_pipeline.framework.param_result import ParamResult
 
 
 class ProblemParamsProvider(ParamsProvider):
@@ -64,3 +64,4 @@ class ProblemParamsProvider(ParamsProvider):
             NotImplementedError: If the subclass does not implement this method.
         """
         raise NotImplementedError("Sub-classes must implement `provide_params`.")
+
