+++ qaoa_training_pipeline/training/recursion.py
@@ -1,4 +1,4 @@
-#
+﻿#
 #
 # (C) Copyright IBM 2024.
 #
@@ -19,7 +19,7 @@ from qiskit.quantum_info import SparsePauliOp
 
 from qaoa_training_pipeline.exceptions import TrainingError
 from qaoa_training_pipeline.training.base_trainer import BaseTrainer
-from qaoa_training_pipeline.training.param_result import ParamResult
+from qaoa_training_pipeline.framework.param_result import ParamResult
 from qaoa_training_pipeline.training.parameter_extenders import PARAMETEREXTENDERS
 from qaoa_training_pipeline.training.scipy_trainer import ScipyTrainer
 
@@ -179,3 +179,4 @@ class RecursionTrainer(BaseTrainer):
         axis.legend()
 
         return fig, axis
+
