+++ qaoa_training_pipeline/training/recursive_transition_states.py
@@ -1,4 +1,4 @@
-"""Recursive transition states trainer."""
+﻿"""Recursive transition states trainer."""
 
 from time import time
 
@@ -10,7 +10,7 @@ from qiskit.quantum_info import SparsePauliOp
 
 from qaoa_training_pipeline.exceptions import TrainingError
 from qaoa_training_pipeline.training.base_trainer import BaseTrainer
-from qaoa_training_pipeline.training.param_result import ParamResult
+from qaoa_training_pipeline.framework.param_result import ParamResult
 from qaoa_training_pipeline.training.scipy_trainer import ScipyTrainer
 from qaoa_training_pipeline.training.transition_states import TransitionStatesTrainer
 
@@ -152,3 +152,4 @@ class RecursiveTransitionStates(BaseTrainer):
         axis.legend()
 
         return fig, axis
+
