+++ qaoa_training_pipeline/training/tqa_trainer.py
@@ -1,4 +1,4 @@
-#
+﻿#
 #
 # (C) Copyright IBM 2024.
 #
@@ -24,7 +24,7 @@ from qaoa_training_pipeline.evaluation.base_evaluator import BaseEvaluator
 from qaoa_training_pipeline.training.base_trainer import BaseTrainer
 from qaoa_training_pipeline.training.functions import BaseAnglesFunction
 from qaoa_training_pipeline.training.history_mixin import HistoryMixin
-from qaoa_training_pipeline.training.param_result import ParamResult
+from qaoa_training_pipeline.framework.param_result import ParamResult
 
 
 class TQATrainerFunction(BaseAnglesFunction):
@@ -325,3 +325,4 @@ class TQATrainer(BaseTrainer, HistoryMixin):
                 raise ValueError("Unknown key in provided train_kwargs.")
 
         return train_kwargs
+
