+++ test/training/test_tqa.py
@@ -1,4 +1,4 @@
-#
+﻿#
 #
 # (C) Copyright IBM 2024.
 #
@@ -11,7 +11,7 @@
 from qiskit.quantum_info import SparsePauliOp
 
 from qaoa_training_pipeline.evaluation.mps_evaluator import MPSEvaluator
-from qaoa_training_pipeline.training.param_result import ParamResult
+from qaoa_training_pipeline.framework.param_result import ParamResult
 from qaoa_training_pipeline.training.tqa_trainer import TQATrainer
 
 # Disable import order for this line. Python has a stdlib test module, but this
@@ -180,3 +180,4 @@ class TestTQA(TrainingPipelineTestCase):
         self.assertTrue(len(result["parameter_history"]) > 0)
         self.assertTrue(len(result["energy_evaluation_time"]) > 0)
         self.assertEqual(trainer.qaoa_angles_function._tqa_schedule.__name__, "lr_schedule")
+
